Deprecate the `packaging` module
