--- docs/api-ref/prefect/packaging/base.md
@@ -1,8 +1,0 @@
----
-description: Prefect Python API for packaging. 
-tags:
-    - Python API
-    - packaging
----
-
-::: prefect.packaging.base
