--- docs/api-ref/prefect/packaging/docker.md
@@ -1,9 +1,0 @@
----
-description: Prefect Python API for packaging with Docker. 
-tags:
-    - Python API
-    - packaging
-    - Docker
----
-
-::: prefect.packaging.docker
