--- docs/api-ref/prefect/packaging/file.md
@@ -1,9 +1,0 @@
----
-description: Prefect Python API for packaging files. 
-tags:
-    - Python API
-    - packaging
-    - file
----
-
-::: prefect.packaging.file
