--- docs/api-ref/prefect/packaging/serializers.md
@@ -1,9 +1,0 @@
----
-description: Prefect Python API for packaging serializers. 
-tags:
-    - Python API
-    - packaging
-    - serializers
----
-
-::: prefect.packaging.serializers
