+++ docs/concepts/filesystems.md
@@ -350,44 +350,3 @@ Prefect provides two abstract file system types, `ReadableFileSystem` and `Write
 - All writeable file systems must implement `write_path` which takes a file path and content and writes the content to the file as bytes. 
 
 A file system may implement both of these types.
-<!-- 
-## Examples
-What does a working example of a MinIO-backed S3 storage block look like within a Prefect flow?
-
-```python
-from prefect import flow, task
-from prefect.deployments import Deployment
-from prefect.filesystems import RemoteFileSystem
-from prefect.logging import get_run_logger
-from prefect.packaging import FilePackager
-
-@task
-def greet_world():
-    logger = get_run_logger()
-    logger.info("Hello world!")
-
-@flow
-def give_greeting() -> str:
-    greet_world()
-    
-minio_packager = FilePackager(
-    filesystem=RemoteFileSystem(
-        basepath="s3://my-bucket",
-        settings={
-            "key": MINIO_ROOT_USER,
-            "secret": MINIO_ROOT_PASSWORD,
-            "client_kwargs": {"endpoint_url": "http://localhost:9000"}
-        }
-    )
-)
-
-Deployment(
-    flow=give_greeting,
-    name="minio_file_package_with_remote_s3fs",
-    packager=minio_packager)
-
-if __name__ == "__main__":
-    give_greeting()
-```
-
-Starting a MinIO server should output logs including login credentials that can be used as your `MINIO_ROOT_USER` and `MINIO_ROOT_PASSWORD`. -->
