+++ mkdocs.yml
@@ -310,11 +310,6 @@ nav:
                       - "highlighters": api-ref/prefect/logging/highlighters.md
                       - "loggers": api-ref/prefect/logging/loggers.md
                 - "prefect.manifests": api-ref/prefect/manifests.md
-                - "prefect.packaging":
-                      - "base": api-ref/prefect/packaging/base.md
-                      - "docker": api-ref/prefect/packaging/docker.md
-                      - "file": api-ref/prefect/packaging/file.md
-                      - "serializers": api-ref/prefect/packaging/serializers.md
                 - "prefect.runner":
                       - "runner": api-ref/prefect/runner/runner.md
                       - "server": api-ref/prefect/runner/server.md
