+++ src/prefect/deprecated/packaging/file.py
@@ -1,5 +1,11 @@
+"""
+DEPRECATION WARNING:
+This module is deprecated as of March 2024 and will not be available after September 2024.
+"""
+
 from uuid import UUID
 
+from prefect._internal.compatibility.deprecated import deprecated_class
 from prefect._internal.pydantic import HAS_PYDANTIC_V2
 
 if HAS_PYDANTIC_V2:
@@ -12,15 +18,22 @@ from typing_extensions import Literal
 from prefect.blocks.core import Block
 from prefect.client.orchestration import PrefectClient
 from prefect.client.utilities import inject_client
+from prefect.deprecated.packaging.base import PackageManifest, Packager, Serializer
+from prefect.deprecated.packaging.serializers import SourceSerializer
 from prefect.filesystems import LocalFileSystem, ReadableFileSystem, WritableFileSystem
 from prefect.flows import Flow
-from prefect.packaging.base import PackageManifest, Packager, Serializer
-from prefect.packaging.serializers import SourceSerializer
 from prefect.settings import PREFECT_HOME
 from prefect.utilities.hashing import stable_hash
 
 
+@deprecated_class(start_date="Mar 2024")
 class FilePackageManifest(PackageManifest):
+    """
+    DEPRECATION WARNING:
+
+    This class is deprecated as of version March 2024 and will not be available after September 2024.
+    """
+
     type: Literal["file"] = "file"
     serializer: Serializer
     key: str
@@ -34,12 +47,17 @@ class FilePackageManifest(PackageManifest):
         return self.serializer.loads(content)
 
 
+@deprecated_class(start_date="Mar 2024")
 class FilePackager(Packager):
     """
+    DEPRECATION WARNING:
+
+    This class is deprecated as of version March 2024 and will not be available after September 2024.
+
     This packager stores the flow as a single file.
 
     By default, the file is the source code of the module the flow is defined in.
-    Alternative serialization modes are available in `prefect.packaging.serializers`.
+    Alternative serialization modes are available in `prefect.deprecated.packaging.serializers`.
     """
 
     type: Literal["file"] = "file"
