+++ src/prefect/deprecated/packaging/orion.py
@@ -1,5 +1,11 @@
+"""
+DEPRECATION WARNING:
+This module is deprecated as of March 2024 and will not be available after September 2024.
+"""
+
 from uuid import UUID
 
+from prefect._internal.compatibility.deprecated import deprecated_class
 from prefect._internal.pydantic import HAS_PYDANTIC_V2
 
 if HAS_PYDANTIC_V2:
@@ -12,12 +18,19 @@ from typing_extensions import Literal
 from prefect.blocks.system import JSON
 from prefect.client.orchestration import PrefectClient
 from prefect.client.utilities import inject_client
+from prefect.deprecated.packaging.base import PackageManifest, Packager, Serializer
+from prefect.deprecated.packaging.serializers import SourceSerializer
 from prefect.flows import Flow
-from prefect.packaging.base import PackageManifest, Packager, Serializer
-from prefect.packaging.serializers import SourceSerializer
 
 
+@deprecated_class(start_date="Mar 2024")
 class OrionPackageManifest(PackageManifest):
+    """
+    DEPRECATION WARNING:
+
+    This class is deprecated as of version March 2024 and will not be available after September 2024.
+    """
+
     type: Literal["orion"] = "orion"
     serializer: Serializer
     block_document_id: UUID
@@ -31,13 +44,18 @@ class OrionPackageManifest(PackageManifest):
         return self.serializer.loads(serialized_flow.encode())
 
 
+@deprecated_class(start_date="Mar 2024")
 class OrionPackager(Packager):
     """
+    DEPRECATION WARNING:
+
+    This class is deprecated as of version March 2024 and will not be available after September 2024.
+
     This packager stores the flow as an anonymous JSON block in the Prefect database.
     The content of the block are encrypted at rest.
 
     By default, the content is the source code of the module the flow is defined in.
-    Alternative serialization modes are available in `prefect.packaging.serializers`.
+    Alternative serialization modes are available in `prefect.deprecated.packaging.serializers`.
     """
 
     type: Literal["orion"] = "orion"
