--- src/prefect/packaging/__init__.py
@@ -1,8 +1,0 @@
-from prefect.packaging.docker import DockerPackager
-from prefect.packaging.file import FilePackager
-from prefect.packaging.orion import OrionPackager
-
-# isort: split
-
-# Register any packaging serializers
-import prefect.packaging.serializers
