+++ tests/deprecated/packaging/test_orion_packager.py
@@ -1,8 +1,8 @@
 import pytest
 
-from prefect.packaging import FilePackager
-from prefect.packaging.file import FilePackageManifest
-from prefect.packaging.serializers import (
+from prefect.deprecated.packaging import OrionPackager
+from prefect.deprecated.packaging.orion import OrionPackageManifest
+from prefect.deprecated.packaging.serializers import (
     ImportSerializer,
     PickleSerializer,
     SourceSerializer,
@@ -15,16 +15,16 @@ from . import howdy
 @pytest.mark.parametrize(
     "serializer", [SourceSerializer(), ImportSerializer(), PickleSerializer()]
 )
-async def test_file_packager_by_serializer(serializer):
-    packager = FilePackager(serializer=serializer)
+async def test_orion_packager_by_serializer(serializer):
+    packager = OrionPackager(serializer=serializer)
     manifest = await packager.package(howdy)
 
-    assert isinstance(manifest, FilePackageManifest)
+    assert isinstance(manifest, OrionPackageManifest)
     unpackaged_howdy = await manifest.unpackage()
     assert unpackaged_howdy("bro") == "howdy, bro!"
 
 
 async def test_packager_sets_manifest_flow_parameter_schema():
-    packager = FilePackager(serializer=SourceSerializer())
+    packager = OrionPackager(serializer=SourceSerializer())
     manifest = await packager.package(howdy)
     assert manifest.flow_parameter_schema == parameter_schema(howdy.fn)
